Provide an argument to return
According to man bash, when return without an argument, the return
status is that of the last command executed in the function body. As we
don't have other commands in the function body, not passing an argument
introduces undefined behaviour. It seems bash uses the return status of
the last command before the function call, which was in our case a falsy
test.